+++ changelog.md
@@ -2,6 +2,12 @@
 
 All notable changes to `TestAssist` will be documented in this file.
 
+## Version 1.1
+
+### Added
+- Assertions for whether a model is inserted or updated in the database.
+- Moved `FabricateModels` to the same `Database` namespace.
+
 ## Version 1.0
 
 ### Added